move hierarchical tests to correct file
